move logic to cogs
