+++ .env.sample
@@ -1,6 +1,7 @@
 DISCORD_SECRET=""
 DISCORD_GUILD_IDS=""
 DISCORD_FORUM_IDS=""
+DISCORD_TESTING_GUILD_ID=""
 DISCORD_STATUS_CHANNEL_ID=""
 DB_PATH="data/mydb.db"
 OPENAI_API_KEY=""
