+++ hordoc/cogs/sync.py
@@ -1,0 +1,56 @@
+from typing import Literal, Optional
+
+import discord
+from discord.ext import commands
+from discord.ext.commands import Greedy, Context
+
+
+class Sync(commands.Cog):
+    """
+    Syncs the command tree.
+    From https://gist.github.com/AbstractUmbra/a9c188797ae194e592efe05fa129c57f
+    """
+
+    def __init__(self, bot):
+        self.bot = bot
+
+    @commands.command()  # type: ignore
+    @commands.guild_only()
+    @commands.is_owner()
+    async def sync(
+        self,
+        ctx: Context,
+        guilds: Greedy[discord.Object],
+        spec: Optional[Literal["~", "*", "^"]] = None,
+    ) -> None:
+        if not guilds:
+            if spec == "~":
+                synced = await ctx.bot.tree.sync(guild=ctx.guild)
+            elif spec == "*":
+                ctx.bot.tree.copy_global_to(guild=ctx.guild)
+                synced = await ctx.bot.tree.sync(guild=ctx.guild)
+            elif spec == "^":
+                ctx.bot.tree.clear_commands(guild=ctx.guild)
+                await ctx.bot.tree.sync(guild=ctx.guild)
+                synced = []
+            else:
+                synced = await ctx.bot.tree.sync()
+
+            from pprint import pprint as pp
+
+            pp(synced)
+            await ctx.send(
+                f"Synced {len(synced)} commands {'globally' if spec is None else 'to the current guild.'}"
+            )
+            return
+
+        ret = 0
+        for guild in guilds:
+            try:
+                await ctx.bot.tree.sync(guild=guild)
+            except discord.HTTPException:
+                pass
+            else:
+                ret += 1
+
+        await ctx.send(f"Synced the tree to {ret}/{len(guilds)}.")
